+++ src/util/FfmpegStreamingProcess.ts
@@ -42,6 +42,7 @@ export class FfmpegStreamingProcess {
     log: PrefixLogger,
     delegate: StreamingDelegate,
     callback?: StreamRequestCallback,
+    startupStartedAt = Date.now(),
   ) {
 
     log.debug(`Stream command: ${videoProcessor} ${ffmpegArgs.map(value => JSON.stringify(value)).join(' ')}`);
@@ -58,11 +59,12 @@ export class FfmpegStreamingProcess {
       if (progress) {
         if (!started && progress.frame > 0) {
           started = true;
-          const runtime = (Date.now() - startTime) / 1000;
-          const message = 'Getting the first frames took ' + runtime + ' seconds.';
-          if (runtime < 5) {
+          const ffmpegRuntime = (Date.now() - startTime) / 1000;
+          const totalRuntime = (Date.now() - startupStartedAt) / 1000;
+          const message = `Getting the first frames took ${totalRuntime} seconds (${ffmpegRuntime} seconds in FFmpeg).`;
+          if (totalRuntime < 5) {
             log.debug(message);
-          } else if (runtime < 22) {
+          } else if (totalRuntime < 22) {
             log.warn(message);
           } else {
             log.error(message);